streamlit/requirements.txt
matplotlib
nltk
numpy
pandas
seaborn
scipy
scikit-learn
glob
re
datetime
tqdm
bs4
python-steam-api
requests
json
func-timeout
selenium
mysql-connector-python